add refresh token

+++ src/api_manager/ApiManager.ts
@@ -8,7 +8,61 @@ const BACKEND_BASE_URL: string = process.env.BACKEND_BASE_URL || 'http://localho
 //const BACKEND_BASE_URL: string = 'https://be-week04-tienpham569us-projects.vercel.app';
 
 export class ApiManager {
-  
+  private static isRefreshing = false;
+  private static refreshSubscribers: Array<(token: string) => void> = [];
+
+  private static onRrefreshed(token: string) {
+    ApiManager.refreshSubscribers.map((callback) => callback(token));
+  }
+
+  private static addRefreshSubscriber(callback: (token: string) => void) {
+    ApiManager.refreshSubscribers.push(callback);
+  }
+
+  private static async refreshToken(refreshToken: string): Promise<string> {
+    const response = await fetch(`${BACKEND_BASE_URL}/user/refreshToken`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({ refreshToken }),
+    });
+
+    if (!response.ok) {
+      throw new Error('Failed to refresh token');
+    }
+
+    const data = await response.json();
+    return data.id_token;
+  }
+
+  private static async fetchWithRefresh(url: string, options: RequestInit, refreshToken: string): Promise<Response> {
+    try {
+      const response = await fetch(url, options);
+      if (response.status === 401) {
+        if (!ApiManager.isRefreshing) {
+          ApiManager.isRefreshing = true;
+          const newToken = await ApiManager.refreshToken(refreshToken);
+          ApiManager.isRefreshing = false;
+          ApiManager.onRrefreshed(newToken);
+        }
+
+        return new Promise((resolve) => {
+          ApiManager.addRefreshSubscriber(async (token: string) => {
+            options.headers = {
+              ...options.headers,
+              'Authorization': `Bearer ${token}`,
+            };
+            resolve(fetch(url, options));
+          });
+        });
+      }
+      return response;
+    } catch (error) {
+      throw error;
+    }
+  }
+
   private mergeHeaders(customHeaders?: any): Headers {
     const defaultHeaders = {
       'accept': 'application/json',
@@ -17,30 +71,51 @@ export class ApiManager {
     return new Headers({ ...defaultHeaders, ...customHeaders });
   }
 
-  static async get(endpoint: string, headers?: any): Promise<any> {
+  static async get(endpoint: string, headers?: any, refreshToken?: string): Promise<any> {
     try {
-      const response = await fetch(`${BACKEND_BASE_URL}/${endpoint}`,{
-        headers: headers,
-      });
+      let response;
+      if (refreshToken) {
+        const options: RequestInit = {
+          method: 'GET',
+          headers: headers,
+        };
+        response = await this.fetchWithRefresh(`${BACKEND_BASE_URL}/${endpoint}`, options, refreshToken);
+      } else {
+        response = await fetch(`${BACKEND_BASE_URL}/${endpoint}`,{
+          headers: headers,
+        });
+      }
+      
       if (!response.ok) {
+        console.log("Response: ", response);
         const errorData = await response.json();
         throw new Error(JSON.stringify(errorData));
       }
       return response.json();
     } catch (error) {
+      console.log("Error: ", error);
       throw error;
     } 
   }
-  static async post(endpoint: string, data: any, headers?: any): Promise<any> {
+  static async post(endpoint: string, data: any, headers?: any, refreshToken?: string): Promise<any> {
     try {
-      const response = await fetch(`${BACKEND_BASE_URL}/${endpoint}`, {
-        method: 'POST',
-        // headers: {
-        //   'Content-Type': 'application/json',
-        // },
-        headers: (headers),
-        body: JSON.stringify(data),
-      });
+      let response;
+      if (refreshToken) {
+        const options: RequestInit = {
+          method: 'POST',
+          headers: headers,
+        };
+        response = await this.fetchWithRefresh(`${BACKEND_BASE_URL}/${endpoint}`, options, refreshToken);
+      } else {
+        response = await fetch(`${BACKEND_BASE_URL}/${endpoint}`, {
+          method: 'POST',
+          // headers: {
+          //   'Content-Type': 'application/json',
+          // },
+          headers: (headers),
+          body: JSON.stringify(data),
+        });
+      }
       if (!response.ok) {
         const errorData = await response.json();
         throw new Error(JSON.stringify(errorData));
@@ -51,13 +126,22 @@ export class ApiManager {
       throw error;
     } 
   }
-  static async put(endpoint: string, data: any, headers?: any): Promise<any> {
+  static async put(endpoint: string, data: any, headers?: any, refreshToken?: string): Promise<any> {
     try {
-      const response = await fetch(`${BACKEND_BASE_URL}/${endpoint}`, {
-        method: 'PUT',
-        headers: headers,
-        body: JSON.stringify(data),
-      });
+      let response;
+      if (refreshToken) {
+        const options: RequestInit = {
+          method: 'PUT',
+          headers: headers,
+        };
+        response = await this.fetchWithRefresh(`${BACKEND_BASE_URL}/${endpoint}`, options, refreshToken);
+      } else {
+        response = await fetch(`${BACKEND_BASE_URL}/${endpoint}`, {
+          method: 'PUT',
+          headers: headers,
+          body: JSON.stringify(data),
+        });
+      }
       if (!response.ok) {
         const errorData = await response.json();
         throw new Error(JSON.stringify(errorData));
@@ -67,12 +151,21 @@ export class ApiManager {
       throw error;
     } 
   }
-  static async delete(endpoint: string, headers?: any): Promise<any> {
+  static async delete(endpoint: string, headers?: any, refreshToken?: string): Promise<any> {
     try {
-      const response = await fetch(`${BACKEND_BASE_URL}/${endpoint}`, {
-        method: 'DELETE',
-        headers: headers,
-      });
+      let response;
+      if (refreshToken) {
+        const options: RequestInit = {
+          method: 'PUT',
+          headers: headers,
+        };
+        response = await this.fetchWithRefresh(`${BACKEND_BASE_URL}/${endpoint}`, options, refreshToken);
+      } else {
+        response = await fetch(`${BACKEND_BASE_URL}/${endpoint}`, {
+          method: 'DELETE',
+          headers: headers,
+        });
+      }
       if (!response.ok) {
         const errorData = await response.json();
         throw new Error(JSON.stringify(errorData));
